feat!: Move identifier to their module
And fix channel id stringification
